refactor(workflow): move vLLM integrations under experimental
